feat(cb2-5716): initial trailer into service letter
