--- src/main/java/uk/gov/dvsa/model/cvs/TrlIntoService.java
@@ -1,53 +1,0 @@
-package uk.gov.dvsa.model.cvs;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-import uk.gov.dvsa.model.Document;
-import uk.gov.dvsa.model.cvs.certificateData.ApplicantDetails;
-
-public class TrlIntoService extends Document {
-
-    @JsonProperty()
-    private ApplicantDetails applicantDetails;
-
-    @JsonProperty()
-    private String vin;
-
-    @JsonProperty()
-    private String typeApproval;
-
-    @JsonProperty()
-    private String trailerId;
-
-    public ApplicantDetails getApplicantDetails() {
-        return applicantDetails;
-    }
-
-    public void setApplicantDetails(ApplicantDetails applicantDetails) {
-        this.applicantDetails = applicantDetails;
-    }
-
-    public String getVin() {
-        return vin;
-    }
-
-    public void setVin(String vin) {
-        this.vin = vin;
-    }
-
-    public String getTypeApproval() {
-        return typeApproval;
-    }
-
-    public void setTypeApproval(String typeApproval) {
-        this.typeApproval = typeApproval;
-    }
-
-    public String getTrailerId() {
-        return trailerId;
-    }
-
-    public void setTrailerId(String trailerId) {
-        this.trailerId = trailerId;
-    }
-
-}
