+++ src/main/resources/assets/stylesheets/trailer-into-service.hbs
@@ -1,0 +1,4 @@
+#letter-subject {
+    font-weight: bold;
+    text-decoration: underline;
+}
