--- src/main/resources/views/CommercialVehicles/TRL_INTO_SERVICE.hbs
@@ -1,10 +1,0 @@
-<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Strict//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-strict.dtd">
-<html>
-    <head>
-        <title>Trailer Into Service</title>
-    </head>
-
-    <body>
-        <p>Hello World</p>
-    </body>
-</html>
